Move `hashCoverCoords` to WirelessBase
